switch to shunit2 from Bats
The current state of Bats [1] doesn't seem clear, even though there is a
community bats-cores [2], the developement certainly isn't very active.

The shunit2 [3] is still quite active, and actually better in some ways.
One of which is that the test function would test all the asserts rather
than when one fails and the rest isn't tested in Bats.

[1] https://github.com/sstephenson/bats
[2] https://github.com/bats-core/bats-core
[3] https://github.com/kward/shunit2